+++ lib/ap/stdio/vfprintf.c
@@ -305,7 +305,7 @@ ocvt_fixed(FILE *f, va_list *args, int flags, int width, int precision,
 
 	if(sgned){
 		if(flags&PTR) snum = (int32_t)va_arg(*args, void *);
-		else if(flags&SHORT) snum = va_arg(*args, int16_t);
+		else if(flags&SHORT) snum = va_arg(*args, int);
 		else if(flags&LONG) snum = va_arg(*args, int32_t);
 		else if(flags&VLONG) snum = va_arg(*args, long long);
 		else snum = va_arg(*args, int);
@@ -321,7 +321,7 @@ ocvt_fixed(FILE *f, va_list *args, int flags, int width, int precision,
 	} else {
 		sign = "";
 		if(flags&PTR) num = (int32_t)va_arg(*args, void *);
-		else if(flags&SHORT) num = va_arg(*args, unsigned short);
+		else if(flags&SHORT) num = va_arg(*args, unsigned int);
 		else if(flags&LONG) num = va_arg(*args, unsigned long);
 		else if(flags&VLONG) num = va_arg(*args, unsigned long long);
 		else num = va_arg(*args, unsigned int);
